% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{impacts}
\alias{impacts}
\title{Impacts estimation of spatial SUR model}
\usage{
impacts(spsurfit, nsim = 1000)
}
\arguments{
\item{nsim}{: Number of simulations. Default=1000}

\item{Form}{: An object of class spsur}
}
\value{
It returns the marginal effects of the estimated Spatial SUR model and simulted values.
  \tabular{ll}{
  \code{table_dir_eff} \tab table of average direct effects. \cr
  \code{table_ind_eff} \tab table of average indirect effects. \cr
  \code{table_tot_eff} \tab table of average total effects. \cr
  \code{sim_dir_eff} \tab simulate direct effects. \cr
  \code{sim_ind_eff} \tab simulate indirect effects. \cr
  \code{sim_tot_eff} \tab simulate total effects. \cr
  }
}
\description{
Returns the marginal effects of a Spatial SUR model (SUR-SAR; SUR-SDM; SUR-SARAR)
}
\details{
The marginal effects (impacts) of spatial autoregressive models (SAR; SDM; SARAR) are more complicated than
usual measurements of impacts for non spatial models. Here we follow LeSage and Pace (2009)
and propose the following summaries for impact measures for Spatial SUR models:\cr
\cr
Average direct effects: The average over all the observations of the effects of the change
of an explanatory variable of a single observation on the choice
probability of that same observation.\cr
\cr
Average indirect effects: The average over all the observations of the effect of a change
 on a explanatory variable on the choice probability of the neighbouring observations.\cr
\cr
Average total effects: The sum of direct and indirect impacts.\cr
}
\examples{

####################################
######## CROSS SECTION DATA ########
####################################

#### Example 1: Spatial Phillips-Curve. Anselin (1988, p. 203)

data(spc)
Tformula <- WAGE83 | WAGE81 ~ UN83 + NMR83 + SMSA | UN80 + NMR80 + SMSA
## A SUR-SAR model
spcSUR.sar <-spsurml(Form=Tformula,data=spc,type="sar",W=Wspc)
summary(spcSUR.sar)
eff.spcSUR.sar <- impacts(spcSUR.sar)

## A SUR-SDM model
spcSUR.sdm <-spsurml(Form=Tformula,data=spc,type="sdm",W=Wspc)
summary(spcSUR.sdm)
eff.spcSUR.sdm <- impacts(spcSUR.sdm)

## A SUR-SARAR model
spcSUR.sarar <-spsurml(Form=Tformula,data=spc,type="sarar",W=Wspc)
summary(spcSUR.sarar)
eff.spcSUR.sarar <- impacts(spcSUR.sarar)

####################################
######## PANEL DATA (nG>1; nT>1) ###
####################################

data(NAT)
Tformula <- HR80  | HR90 ~ PS80 + UE80 | PS90 + UE90
## A SUR-SAR model
NATSUR.sar <-spsurml(Form=Tformula,data=NAT,type="sar",W=W)
summary(NATSUR.sar)
eff.NAT.sar <- impacts(NATSUR.sar,nsim=10)

####################################
######## PANEL DATA (nG=1; nT>1) ###
####################################

data("unemp_it_short")
data("W_italy")
form_un <- unrate  ~ empgrowth + partrate + agri + cons + serv
unempitml_sar <- spsurtime(Form=form_un,data=unemp_it,time=unemp_it$year,W=W_italy,type="sar",method="ml")
summary(unempitml_sar)

eff.unempitml_sar <- impacts(unempitml_sar,nsim=100)

}
\references{
Mur, J., Lopez, F., and Herrera, M. (2010). Testing for spatial effects in seemingly unrelated regressions. \emph{Spatial Economic Analysis}, 5(4), 399-440.
\cr
\cr
Lopez, F.A., Mur, J., and Angulo, A. (2014). Spatial model selection strategies in a SUR framework. The case of regional productivity in EU. \emph{Annals of Regional Science}, 53(1), 197-220.
\cr
\cr
LeSage, J., and Pace, R. K. (2009). Introduction to spatial econometrics. Chapman and Hall/CRC.
}
\seealso{
\code{\link{spsurml}},  \code{\link{spsur3sls}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Jesús Mur  \tab \email{jmur@unizar.es} \cr
  }
}
