% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_betas_spsur.R, R/lrtestspsur.R
\name{lrtestspsur}
\alias{lrtestspsur}
\alias{lr_betas_spsur}
\alias{lrtestspsur}
\title{Likelihood ratio test}
\usage{
lr_betas_spsur(Form = NULL, data = NULL, R = NULL, r = NULL, W = NULL,
  time = NULL, X = NULL, Y = NULL, nG = NULL, nR = NULL, nT = NULL,
  p = NULL, type = "sim", printmodels = FALSE, cov = FALSE,
  trace = FALSE)

lrtestspsur(Form = NULL, data = NULL, W = NULL, X = NULL, Y = NULL,
  time = NULL, nG = NULL, nR = NULL, nT = NULL)
}
\arguments{
\item{W}{: A RxR spatial weight matrix.}

\item{X}{: Data matrix R*TxK (K is a vector ...)}

\item{Y}{: Data vector R*Tx1 (firs space, second time)}

\item{nG}{: number of equations}

\item{nR}{: number of spatial observations}

\item{nT}{: Number of time periods}

\item{nT}{: Number of time periods}

\item{nG}{: number of equations}

\item{nR}{: number of spatial observations}

\item{Y}{: Data vector R*Tx1 (firs space, second time)}

\item{X}{: Data matrix R*TxK (K is a vector ...)}

\item{W}{: A RxR spatial weight matrix.}
}
\value{
The LR tests

The LR tests
}
\description{
The function report serveral LR tests

The function report serveral LR tests
}
\details{
The SUR models
Various tests of spatial autocorrelation are obtain based on the principle of the Lagrange Multiplier in a maximum-likelihood framework.

W is a spatial weights matrix

The SUR models
Various tests of spatial autocorrelation are obtain based on the principle of the Lagrange Multiplier in a maximum-likelihood framework.

W is a spatial weights matrix
}
\examples{

data(Sar)
nT <- 4 # Number of periods
nG <- 3 # Number equations
nR <- ncol(Ws)
lrtestspsur(W=Ws,X=XXsar,Y=Ysar,nG=nG,nR=nR,nT=nT)
examples
data(spc)
Tformula <- WAGE83 | WAGE81 ~ UN83 + NMR83 + SMSA | UN80 + NMR80 + SMSA
H0: equal  beta for SMSA in both equations.
R1 <- matrix(c(0,0,0,1,0,0,0,-1),nrow=1)
r1 <- matrix(0,ncol=1)
lr_smsa <-  lr_betas_spsur(Form=Tformula,data=spc,
                              W=Wspc,type="sar",R=R1,r=r1,
                              trace=TRUE,printmodels = TRUE)
####### Example with spatio-temporal SUR. 
####### Assumption: nR,nT > 1 and nG = 1. Database is a spatio-temporal panel
data("unemp_it_short") 
data("W_italy")
form_un <- unrate  ~ empgrowth + partrate + agri + cons + serv
H0: equal emprgrowth beta in equations 1, 3, and 4 
R <- matrix(0,nrow=2,ncol=30)
R[1,2] <- 1; R[1,14] <- -1
R[2,2] <- 1; R[2,20] <- -1
r <- matrix(0,nrow=2,ncol=1)
lr_partrate <-  lr_betas_spsur(Form=form_un,data=unemp_it,time=unemp_it$year,
                          W=W_italy,type="sar",R2=R,r=r2,
                         trace=TRUE,printmodels = TRUE)
lr_betas_spcSUR_sar <- lr_betas_spsur(Form=Tformula,data=spc,
                                 R=R1,r=r1,type="sar",W=Wspc,
                                 printmodels=TRUE)

data(Sar)
nT <- 4 # Number of periods
nG <- 3 # Number equations
nR <- ncol(Ws)
lrtestspsur(W=Ws,X=XXsar,Y=Ysar,nG=nG,nR=nR,nT=nT)
########################################################################
##########     TEMPORAL CORRELATIONS (nG = 1 and nT > 1)     ###########
########################################################################
data("unemp_it_short") 
data("W_italy")
form_un <- unrate  ~ empgrowth + partrate + agri + cons + serv
lr_time <- lrtestspsur(Form=form_un,data=unemp_it,time=unemp_it$year,W=W_italy)

}
\references{
López, F.A., Mur, J., & Angulo, A. (2014). Spatial model selection strategies in a SUR framework. The case of regional productivity in EU. The Annals of Regional Science, 53(1), 197-220.

López, F.A., Mur, J., & Angulo, A. (2014). Spatial model selection strategies in a SUR framework. The case of regional productivity in EU. The Annals of Regional Science, 53(1), 197-220.
}
\seealso{
\code{\link{spsurml}}

\code{\link{spsurml}}
}
