% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_spsur.R
\name{methods_spsur}
\alias{methods_spsur}
\alias{anova}
\alias{anova.spsur}
\alias{coef}
\alias{coef.spsur}
\alias{fitted}
\alias{fitted.spsur}
\alias{logLik}
\alias{logLik.spsur}
\alias{print}
\alias{print.spsur}
\alias{residuals}
\alias{residuals.spsur}
\alias{vcov}
\alias{vcov.spsur}
\title{Methods for class spsur}
\usage{
\method{anova}{spsur}(object, ...)

\method{coef}{spsur}(object, ...)

\method{fitted}{spsur}(object, ...)

\method{logLik}{spsur}(object, ...)

\method{print}{spsur}(x, ...)

\method{residuals}{spsur}(object, ...)

\method{vcov}{spsur}(object, ...)
}
\arguments{
\item{object}{An \code{spsur} object created by \code{\link{spsurml}},
\code{\link{spsur3sls}} or \code{\link{spsurtime}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Same than *object* for `print` method.}
}
\description{
Methods for class spsur
}
