% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.spsur.R
\name{print.summary.spsur}
\alias{print.summary.spsur}
\title{Print method for objects of class summary.spsur.}
\usage{
\method{print}{summary.spsur}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \emph{summary.spsur}.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for objects of class summary.spsur.
}
\examples{
 # See examples for \\code{\\link{spsurml}} or
 # \\code{\\link{spsur3sls}} functions.
}
\seealso{
\code{\link{summary.spsur}}.
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Jesús Mur  \tab \email{jmur@unizar.es} \cr
  }
}
