% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spc}
\alias{spc}
\title{A classical Spatial Phillips-Curve}
\format{
A data frame with 25 rows and 10 variables:
\describe{
  \item{COUNTY}{County coded as a name.}
  \item{WAGE83}{Changes in wage rates for 1983.}
  \item{UN83}{Inverse unemployment rate in 1983.}
  \item{NMR83}{Net migration rate 1983.}
  \item{SMSA}{Dummy variable to identify counties defined as
    Standard Metropolitan Statistical Areas (SMSA = 1).}
  \item{WAGE82}{Changes in wage rates for 1982.}
  \item{WAGE81}{Changes in wage rates for 1981.}
  \item{UN80}{Inverse unemployment rate in 1980.}
  \item{NMR80}{Net migration rate 1983.}
  \item{WAGE80}{changes in wage rates.}
}
}
\source{
Anselin (1988, p. 203-211)
}
\usage{
spc
}
\description{
A data set from Anselin (1988, p. 203-2011) used to estimate a Spatial
Phillips-Curve for 25 counties in South-West Ohio for two time
periods (1981 and 1983).
}
\references{
\itemize{
    \item Anselin, L. (1988). \emph{Spatial Econometrics:
      Methods and Models}. Springer Science & Business Media.
  }
}
\keyword{datasets}
