% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsur3sls.R
\name{spsur3sls}
\alias{spsur3sls}
\title{3sls estimation of spatial SUR model including restrictions of beta}
\usage{
spsur3sls(Form = NULL, data = NULL, R = NULL, r = NULL, W = NULL,
  X = NULL, Y = NULL, nG = NULL, nR = NULL, nT = NULL, p = NULL,
  demean = FALSE, type = "sar", maxlagW = 2, trace = TRUE)
}
\arguments{
\item{Form}{: An object create with \code{\link[Formula]{Formula}}. package allowing for multiple responses and multiple parts of regressors}

\item{data}{: An object of class data.frame or a matrix}

\item{R}{: Default NULL. A matrix including coefficients of linear hypothesis on beta}

\item{r}{: Default NULL. A vector including independent vector of linear hypothesis on beta}

\item{W}{: A RxR spatial weight matrix}

\item{X}{: Default NULL. Data matrix nRxnTxp (p=sum(p_g) where p_g is the number of independent variables for g-th equation, g=1,...,nG)}

\item{Y}{: Default NULL. Data vector nRxnTx1 (firs: space dimension | second: time periods)}

\item{nG}{: number of equations}

\item{nR}{: number of spatial observations}

\item{nT}{: Number of time periods}

\item{p}{: Number of regressors by equation (p=sum(p_g) where p_g is the number of independent variables for g-th equation, g=1,...,nG)}

\item{demean}{: Demeaning the panel. Default = FALSE}

\item{type}{: type of model "sar" or "sdm"}

\item{maxlagW}{: maximum order of the instruments. Default=2}

\item{trace}{: TRUE to see intermediate results (TRUE or FALSE). Defalult is TRUE}
}
\value{
Results of ML estimation of a SUR model with spatial effects. A list with:
  \tabular{ll}{
  \code{type}   \tab the matched call. \cr
  \code{call}   \tab the model estimate. \cr
  \code{betas} \tab Coefficients \beta estimated by 3sls. \cr
  \code{deltas} \tab Spatial autocorrelation coefficients estimated by 3sls. \cr
  \code{betas_sd} \tab Standart desviation of spatial autocorrelation coefficients estimated by 3sls. \cr
  \code{deltas_sd} \tab Spatial autocorrelation coefficients estimated by 3sls. \cr
  \code{R2}    \tab Determination coefficient between the explained variable and its predictions. Global and for each equation. \cr
  \code{Sigma}    \tab estimated covariance matrix of residuals between equations. \cr
  \code{Sigma_corr}    \tab estimated correlation matrix of residuals between equations. \cr
  \code{residuals}    \tab residuals of model. \cr
  \code{fitted.values}    \tab fitted values of model. \cr
  }
}
\description{
The function fits by three-stage least squares estimation of spatial a
spatial SUR-lag spatial or SUR-SDM model.
}
\details{
Three-stage least squate (3sls) estimation of the SUR models by using spatially lagged X variables as instruments for the spatially lagged dependent variable: \cr
\cr
SUR-SAR: Spatial autorregresive model\cr
\deqn{y_{gt} = \lambda_{g} Wy_{gt} + X_{gt} \beta_{g} + \epsilon_{gt}}
SUR-SDM: Spatial Durbin model\cr
\deqn{y_{gt} = \lambda_{g} Wy_{gt} + X_{gt} \beta_{g} + WX_{gt} \theta_{g} + \epsilon_{gt}}
where y_{gt}, u_{gt} and \eqn{\epsilon_{gt}} are (nRx1) vectors; X_{gt} is a matrix of exogenous variables of
order (nRxsum(p)); \eqn{\lambda_g} and \eqn{\rho_g} are parametres of spatial dependence; W is a nRxnR matrix of spatial interactions.

The instruments used are WX and \eqn{W^{2}}X by default\cr
By default an object create with \code{\link[Formula]{Formula}} is the input of this function. Alternatively a Y vector of dependent
variables and X matrix with independent varaibles could be include as imputs obtain by exaple with \code{\link{dgp_spSUR}}\cr
\cr
The marginal Multiplier tests are used to test for no correlation in one part of the model allowing for spatial correlation in the other. \cr
The LM(LM(\eqn{\lambda}|\eqn{\rho})) is the test for sustantive spatial autocorrelation in a model with spatial autocorrelation in error term.
The LM(\eqn{\rho}|\eqn{\lambda}) is the test for spatial error correlation in a model with sustanative spatial correlation.
}
\examples{

#################################################
######## CROSS SECTION DATA (nG>1; nT=1) ########
#################################################

#### Example 1: Spatial Phillips-Curve. Anselin (1988, p. 203)
## A SUR model without spatial effects
data(spc)
Tformula <- WAGE83 | WAGE81 ~ UN83 + NMR83 + SMSA | UN80 + NMR80 + SMSA

## A SUR-SAR model (3SLS Estimation)
spcSUR.sar.3sls <-spsur3sls(Form=Tformula,data=spc,type="sar",W=Wspc)
summary(spcSUR.sar.3sls)

## A SUR-SDM model (3SLS Estimation)
spcSUR.sdm.3sls <-spsur3sls(Form=Tformula,data=spc,type="sdm",W=Wspc)
summary(spcSUR.sdm.3sls)

####################################
######## PANEL DATA (nG>1; nT>1) ###
####################################

data(Sar)
nT <- 4 # Number of periods
nG <- 3 # Number equations
nR <- 49 # Number of spatial units
SUR.sar.3sls <-spsur3sls(Y=Ysar,X=XXsar,nG=nG,nR=nR,nT=nT,p=c(5,5,5),W=Ws,type="sar")
summary(SUR.sar.3sls)
################### Durbin case with demeaning in nT
   SUR.sdm.3sls.dem <-spsur3sls(Y=Ysar,X=XXsar,nG=nG,nR=nR,nT=nT,p=c(5,5,5),W=Ws,
                            type="sdm",demean=TRUE)
   summary(SUR.sdm.3sls.dem)

#### Estimation with demeaning in nT (intercept of each equation dissapears)
SUR.sar.3sls.dem <-spsur3sls(Y=Ysar,X=XXsar,nG=nG,nR=nR,nT=nT,p=c(5,5,5),W=Ws,
                     type="sar",demean=TRUE)



#' #### Example 2: Homicides + Socio-Economics (1960-90)
Homicides and selected socio-economic characteristics for continental U.S. counties. Data for four decennial census years: 1960, 1970, 1980 and 1990.
https://geodacenter.github.io/data-and-lab/ncovr/
data(NAT)
Tformula <- HR80  | HR90 ~ PS80 + UE80 | PS90 + UE90

## A SUR-SAR model
NATSUR.sar.3sls <-spsur3sls(Form=Tformula,data=NAT,type="sar",W=W,maxlagW=2)
summary(NATSUR.sar.3sls)


}
\references{
Mur, J., López, F., & Herrera, M. (2010). Testing for spatial effects in seemingly unrelated regressions. \emph{Spatial Economic Analysis}, 5(4), 399-440.
\cr
\cr
López, F.A., Mur, J., & Angulo, A. (2014). Spatial model selection strategies in a SUR framework. The case of regional productivity in EU. \emph{Annals of Regional Science}, 53(1), 197-220.
\cr
\cr
}
\seealso{
\code{\link{lmtestspsur}}, \code{\link{spsurml}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Jesus Mur  \tab \email{jmur@unizar.es} \cr
  }
}
