% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsurml.R
\name{spsurml}
\alias{spsurml}
\title{Maximum likelihood estimation of spatial SUR model.}
\usage{
spsurml(Form = NULL, data = NULL, R = NULL, b = NULL, W = NULL,
  X = NULL, Y = NULL, G = NULL, N = NULL, Tm = NULL, p = NULL,
  demean = FALSE, type = "sim", cov = TRUE, control = list(tol =
  0.05, maxit = 200, trace = TRUE))
}
\arguments{
\item{Form}{An object created with the package \code{\link[Formula]{Formula}}
that describes the model to be estimated. This model may contain several
responses (explained variables) and a varying number of regressors in each equation.}

\item{data}{An object of class data.frame or a matrix.}

\item{R}{A row vector of order \emph{(1xpr)} with  the set of \emph{r} linear constraints
on the \emph{beta} parameters. The \emph{first} restriction appears in the first \emph{p} terms,
the second restriction in the next \emph{p} terms and so on. Default = \code{NULL}.}

\item{b}{A column vector of order \emph{(rx1)} with the values of the linear restrictions on the
\emph{beta} parameters. Default = \code{NULL}.}

\item{W}{A spatial weighting matrix of order \emph{(NxN)}, assumed to be the
same for all equations and time periods.}

\item{X}{A data matrix of order \emph{(NTmGxp)} with the observations of the regressors
The number of covariates in the SUR model is p = \eqn{sum(p_{g})} where \emph{\eqn{p_{g}}}
is the number of regressors (including the intercept) in the g-th equation, \emph{g = 1,...,G}).
The specification of \emph{X} is only necessary if not available a \code{\link[Formula]{Formula}}
and a data frame. Default = \code{NULL}.}

\item{Y}{A column vector of order \emph{(NTmGx1)}, with the observations of the
explained variables. The ordering of the data must be (first) equation,
(second) time dimension and (third) Cross-sectional/spatial units.
The specification of \emph{Y} is only necessary if not available a \code{\link[Formula]{Formula}}
and a data frame. Default = \code{NULL}.}

\item{G}{Number of equations.}

\item{N}{Number of cross-section or spatial units}

\item{Tm}{Number of time periods.}

\item{p}{Number of regressors by equation, including the intercept. \emph{p} can be a row
vector of order \emph{(1xG)}, if the number of regressors is not the same for all the
equations, or a scalar, if the \emph{G} equations have the same number of regressors.
The specification of \emph{p} is only necessary if not available a \code{\link[Formula]{Formula}}
and a data frame.}

\item{demean}{Logical value to allow for the demeaning of panel data. In this case,
\code{\link{spsurml}} substracts the individual mean to each spatial or cross-sectional
unit. Default = \code{FALSE}.}

\item{type}{Type of spatial model specification: \strong{"sim"},\strong{"slx"}, \strong{"slm"},
\strong{"sem"}, \strong{"sdm"}, \strong{"sdem"} or \strong{"sarar"}. Default = \code{"sim"}.}

\item{cov}{Logical value to show the covariance matrix of the \emph{beta} coefficients.
Default = \code{TRUE}.}

\item{control}{List of additional control arguments.}
}
\value{
Output of the  maximum-likelihood estimation of the specified spatial SUR model. A list with:
  \tabular{ll}{
    \code{call} \tab Matched call. \cr
    \code{type} \tab  Type of model specified. \cr
    \code{betas} \tab Estimated coefficients for the regressors. \cr
    \code{deltas} \tab Estimated spatial coefficients. \cr
    \code{se_betas} \tab Estimated standard errors for the estimates of \emph{beta}. \cr
    \code{se_deltas} \tab Estimated standard errors for the estimates of the spatial coefficients. \cr
    \code{cov} \tab Estimated covariance matrix for the estimates of \emph{beta's} and spatial coefficients.\cr
    \code{llsur} \tab Value of the likelihood function at the maximum-likelihood estimates. \cr
    \code{R2} \tab Coefficient of determination for each equation, obtained as the squared
     of the correlation coefficient between the corresponding explained variable and
      its estimate. \emph{spsurml} also shows a \emph{global} coefficient of
       determination obtained, in the same manner, for the set of \emph{G} equations. \cr
    \code{Sigma} \tab Estimated covariance matrix for the residuals of the \emph{G} equations. \cr
    \code{Sigma_corr} \tab Estimated correlation matrix for the residuals of the \emph{G} equations. \cr
    \code{Sigma_inv} \tab Inverse of \code{Sigma}, the \emph{(GxG)} covariance matrix of
     the residuals of the SUR model. \cr
    \code{residuals} \tab Residuals of the model. \cr
    \code{df.residuals} \tab Degrees of freedom for the residuals. \cr
    \code{fitted.values} \tab Estimated values for the dependent variables. \cr
    \code{BP} \tab Value of the Breusch-Pagan statistic to test the null hypothesis
     of diagonality among the errors of the \emph{G} equations. \cr
    \code{LMM} \tab Marginal Lagrange Multipliers, LM(\eqn{\rho}|\eqn{\lambda}) and
      LM(\eqn{\lambda}|\eqn{\rho}), to test for omitted spatial effects in the specification. \cr
    \code{G} \tab Number of equations. \cr
    \code{N} \tab Number of cross-sections or spatial units. \cr
    \code{Tm} \tab Number of time periods. \cr
    \code{p} \tab Number of regressors by equation (including intercepts). \cr
    \code{demean} \tab Logical value used for demeaning. \cr
    \code{Y} \tab Vector \emph{Y} of the explained variables of the SUR model. \cr
    \code{X} \tab Matrix \emph{X} of the regressors of the SUR model. \cr
 }
}
\description{
This function estimates spatial SUR models using maximum-likelihood methods.
 The number of equations, time periods and cross-sectional units is not restricted.
 The user can choose between different spatial specifications as described below.
 The estimation procedure allows for the introduction of linear restrictions
 on the \eqn{\beta} parameters associated to the regressors.
}
\details{
The list of (spatial) models that can be estimated with the \emph{spsurml} function are:
 \itemize{
    \item \strong{"sim"}: SUR model with no spatial effects
      \deqn{ y_{tg} = X_{tg} \beta_{g} + \epsilon_{tg} }
    \item \strong{"slx"}: SUR model with spatial lags of the regressors
      \deqn{ y_{tg} = X_{tg} \beta_{g} + WX_{tg} \theta_{g} + \epsilon_{tg} }
    \item \strong{"slm"}: SUR model with spatial lags of the explained variables
      \deqn{y_{tg} = \lambda_{g} Wy_{tg} + X_{tg} \beta_{g} + \epsilon_{tg} }
    \item \strong{"sem"}: SUR model with spatial errors
      \deqn{ y_{tg} = X_{tg} \beta_{g} + u_{tg} }
      \deqn{ u_{tg} = \rho_{g} Wu_{tg} + \epsilon_{tg} }
    \item \strong{"sdm"}: SUR model of the Spatial Durbin type
      \deqn{ y_{tg} = \lambda_{g} Wy_{tg} + X_{tt} \beta_{g} + WX_{tg} \theta_{g} + \epsilon_{tg} }
    \item \strong{"sdem"}: SUR model with spatial lags of the regressors and spatial errors
      \deqn{ y_{tg} = X_{tg} \beta_{g} + WX_{tg} \theta_{g} + u_{tg} }
      \deqn{ u_{tg} = \rho_{g} W u_{tg} + \epsilon_{tg} }
    \item \strong{"sarar"}: SUR model with spatial lags of the explained variables and spatial
      errors
      \deqn{ y_{tg} = \lambda_{g} Wy_{tg} + X_{tg} \beta_{g} + u_{tg} }
      \deqn{ u_{tg} = \rho_{g} W u_{tg} + \epsilon_{tg} }
  }
}
\section{Control arguments}{

  \tabular{ll}{
    \code{tol} \tab Numerical value for the tolerance for the estimation algorithm until
     convergence. Default = 1e-3. \cr
    \code{maxit} \tab Maximum number of iterations until convergence; it must be an integer
     value. Default = 200. \cr
    \code{trace} \tab A logical value to show intermediate results during the estimation process.
      Default = \emph{TRUE}. \cr
}
}

\examples{

#################################################
######## CROSS SECTION DATA (G>1; Tm=1) ########
#################################################

#### Example 1: Spatial Phillips-Curve. Anselin (1988, p. 203)
## It usually requires 2-3 minutes maximum...
rm(list = ls()) # Clean memory
data(spc)
Tformula <- WAGE83 | WAGE81 ~ UN83 + NMR83 + SMSA | UN80 + NMR80 + SMSA
spcsur.sim <-spsurml(Form = Tformula, data = spc, type = "sim", W = Wspc)
summary(spcsur.sim)

## A SUR-SLX model
spcsur.slx <-spsurml(Form = Tformula, data = spc, type = "slx", W = Wspc)
summary(spcsur.slx)
\dontrun{
## Not run:
## A SUR-SLM model
spcsur.slm <-spsurml(Form = Tformula, data = spc, type = "slm", W = Wspc)
summary(spcsur.slm)
rm(spcsur.slm) # remove
}
}
\references{
\itemize{
    \item Mur, J., López, F., and Herrera, M. (2010). Testing for spatial
      effects in seemingly unrelated regressions.
      \emph{Spatial Economic Analysis}, 5(4), 399-440.
     \item López, F.A., Mur, J., and Angulo, A. (2014). Spatial model
       selection strategies in a SUR framework. The case of regional
       productivity in EU. \emph{Annals of Regional Science}, 53(1), 197-220.
     \item Breusch T, Pagan A (1980) The Lagrange multiplier test and its
       applications to model specification in econometrics. \emph{Rev Econ Stud} 47: 239-254
  }
}
\seealso{
\code{\link{spsur3sls}}, \code{\link{lmtestspsur}}, \code{\link{wald_betas}}, \code{\link{lrtestspsur}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Jesús Mur  \tab \email{jmur@unizar.es} \cr
  }
}
