% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{W}
\alias{W}
\title{Spatial weight matrix for U.S. Counties}
\format{A row-standardized squared matrix with 3085 rows and columns.
The rows and columns follow the same order than Counties included in
\code{\link{NCOVR}} data frame.}
\source{
\url{https://geodacenter.github.io/data-and-lab/ncovr/}
}
\usage{
W
}
\description{
A spatial weight matrix row-standardized based on first order
  contiguity criterium.
}
\references{
\itemize{
    \item Baller, R., L. Anselin, S. Messner, G. Deane and D. Hawkins
      (2001). Structural covariates of US county homicide rates:
      incorporating spatial effects. \emph{Criminology} 39, 561-590.
}
}
\keyword{datasets}
