% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmtestspsur.R
\name{lmtestspsur}
\alias{lmtestspsur}
\title{Testing for spatial effects in Seemingly Unrelated Regression}
\usage{
lmtestspsur(Form = NULL, data = NULL, W = NULL, X = NULL, Y = NULL,
  time = NULL, nG = NULL, nR = NULL, nT = NULL, print_table = TRUE)
}
\arguments{
\item{Form}{: An object create with \code{\link[Formula]{Formula}}. package allowing for multiple responses and multiple parts of regressors}

\item{W}{: A nRxnR spatial weight matrix.}

\item{X}{: Default NULL. Data matrix nRxnTxp (p=sum(p_g) where p_g is the number of independent variables for g-th equation, g=1,...,nG)}

\item{Y}{: Default NULL. Data vector nRxnTx1 (firs: space dimension | second: time periods)}

\item{nG}{: Number of equations}

\item{nR}{: Number of spatial observations}

\item{nT}{: Number of time periods}

\item{print_table}{: FALSE or TRUE to print the output. Default = TRUE}
}
\value{
A list, each with
  \tabular{ll}{
  \code{stat_names}   \tab the name of Lagrange Multiplier tests \cr
  \code{stat}   \tab the value of five Lagrange Multiplier tests \cr
  \code{df} \tab number of degrees of freedom for each test \cr
  }
}
\description{
The function \code{\link{spsurml}} reports five statistis for testing spatial dependence in Semmingly Unrelated Regression models (SUR).
The five statistics are based on Lagrangian Multipliers principe. The tests can be used with one cross sectional data set or in a panel data
framework.\cr
The statistics are similar to the LM test of spatial autocorrelation for a simpgle equation (see \code{\link[spdep]{lm.LMtests}} in package \link[pkg]{spdep}).
The LM test for error dependence (LM-SUR-err), the LM test for a missing spatially lagged dependent variable (LM-SUR-lag) and LM-SUR-SARAR for simultaneus
spatial lagged dependent variable and spatial error structure.
Variants of these robust to the presence of the other (RLM-SUR-err, RLM-SUR-lag) are implement too.
}
\details{
The model specification in the general case of nR individuals, nT cross sections and nG equations, with spatial interaction mechanisms, as follows:
\deqn{y_{gt}=\lambda_{g} W y_{gt} + X_{gt} \beta_{g} + u_{gt}}
\deqn{u_{gt} = \rho_{g} W u_{gt} + \epsilon_{gt}}
where y_{gt}, u_{gt} and \eqn{\epsilon_{gt}} are (nRx1) vectors; X_gt is a matrix of exogenous variables of order (nRxKg); \eqn{\lambda_g} and \eqn{\rho_g} are parametres of spatial dependence; W is the RxR matrix.

Five tests of diagnotic of spatial autocorrelation based on the principle of the Lagrange Multiplier is develop in this code:\cr
   LM-SUR-lag   : \eqn{H_0: \lambda=0 (\forall g) versus H_A: No H_0}\cr
   LM-SUR-err   : \eqn{H_0: \rho=0 (\forall g) versus H_A: No H_0}\cr
   LM-SUR-sarar : \eqn{H_0: \lambda=0 and \rho=0 (\forall g) versus H_A: No H_0}\cr

Two robust tests are included under local misspecifications of the alternative. The LM-SUR-lag and LM-SUR-err can be strongly oversized\cr
For cross-section (nT=1) data and nG(>1) equations
}
\examples{
#################################################
######## CROSS SECTION DATA (nG>1; nT=1) ########
#################################################

#### Example 1: Spatial Phillips-Curve. Anselin (1988, p. 203)
data("spc")
Tformula <- WAGE83 | WAGE81 ~ UN83 + NMR83 + SMSA | UN80 + NMR80 + SMSA
LMs <- lmtestspsur(Form=Tformula,data=spc,W=Wspc)

#################################################
######## PANEL DATA (nG>1; nT>1)         ########
#################################################

#### Example 2: Homicides & Socio-Economics (1960-90)
Homicides and selected socio-economic characteristics for continental U.S. counties. Data for four decennial census years: 1960, 1970, 1980 and 1990.
https://geodacenter.github.io/data-and-lab/ncovr/
data("NAT")
Tformula <- HR70 | HR80  | HR90 ~ PS70 + UE70 |PS80 + UE80 | PS90 + UE90
LMs <- lmtestspsur(Form=Tformula,data=NAT,W=W)

## With different number of exogenous variables in each equation
Tformula <- HR70 | HR80  | HR90 ~ PS70 + UE70 | PS80 + UE80 +RD80 | PS90 + UE90 + RD90 + PO90
LMs <- lmtestspsur(Form=Tformula,data=NAT,W=W)

################################################################
######## PANEL DATA: TEMPORAL CORRELATIONS (nG=1; nT>1) ########
################################################################

#### Example 3: Unemployment data in Italy
data("unemp_it_short")
data("W_italy")
form_un <- unrate  ~ empgrowth + partrate + agri + cons + serv
LM_time <- lmtestspsur(Form=form_un,data=unemp_it,time=unemp_it$year,W=W_italy)

}
\references{
Mur, J., Lopez, F., & Herrera, M. (2010). Testing for spatial effects in seemingly unrelated regressions. \emph{Spatial Economic Analysis}, 5(4), 399-440.
\cr
\cr
Lopez, F.A., Mur, J., & Angulo, A. (2014). Spatial model selection strategies in a SUR framework. The case of regional productivity in EU. \emph{Annals of Regional Science}, 53(1), 197-220.
}
\seealso{
\code{\link{spsurml}},\code{\link{spsur3sls}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{Roman.Minguez@uclm.es} \cr
  Jesus Mur  \tab \email{jmur@unizar.es}\cr
 }
}
