% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsurml.R
\name{spsurml}
\alias{spsurml}
\title{Maximum likelihood estimation of spatial SUR model}
\usage{
spsurml(Form = NULL, data = NULL, R = NULL, r = NULL, W = NULL,
  X = NULL, Y = NULL, nG = NULL, nR = NULL, nT = NULL, p = NULL,
  demean = FALSE, type = "sim", cov = TRUE, trace = TRUE)
}
\arguments{
\item{Form}{: An object create with \code{\link[Formula]{Formula}} package allowing for multiple responses and multiple parts of regressors}

\item{data}{: An object of class data.frame or a matrix}

\item{R}{: Default NULL. A matrix including coefficients of linear hypothesis on beta}

\item{r}{: Default NULL. A vector including independent vector of linear hypothesis on beta}

\item{W}{: A RxR spatial weight matrix}

\item{X}{: Default NULL. Data matrix nRxnTxp (p=sum(p_g) where p_g is the number of independent variables for g-th equation, g=1,...,nG)}

\item{Y}{: Default NULL. Data vector nRxnTx1 (firs: space dimension | second: time periods)}

\item{nG}{: number of equations}

\item{nR}{: number of spatial observations}

\item{nT}{: Number of time periods}

\item{p}{: Number of regressors by equation (including independent term). A number (resp. vector) if equal (resp. no equal) regressors by equation}

\item{demean}{: Default FALSE.}

\item{type}{: type of model "sim" "slx" "sar" "sem" "sdm" "sdem" "sarar" . Defalult is "sim"}

\item{cov}{: get the covariance matrix (TRUE or FALSE). Defalult is TRUE}

\item{trace}{: To get intermediate results (TRUE or FALSE). Defalult is TRUE}
}
\value{
Results of ML estimation of a SUR model with spatial effects. A list with:
  \tabular{ll}{
  \code{betas} \tab Coefficients estimated by maximum likelihood. \cr
  \code{llsur}    \tab the maximum value of likelihood function. \cr
  \code{Sigma}    \tab estimated covariance matrix of residuals between equations. \cr
  \code{Sigma_corr}    \tab estimated correlation matrix of residuals between equations. \cr
  \code{residuals}    \tab residual of model. \cr
  \code{fitted.values}    \tab fitted values of model. \cr
  \code{call}   \tab the model estimate. \cr
  \code{type}   \tab the matched call. \cr
  \code{se_betas}   \tab Standard desviation of beta coeeffients. \cr
  \code{BP}   \tab Breusch-Pagan test of diagonality of the SUR matrix of covariance residuals. \cr
  \code{R2}    \tab Determination coefficient between the explained varaible and its predictionsthe p-value of coefficientes. Global and for each equation. \cr
  \code{BP}    \tab Breush-Pagan test (1980) of diagonaltiy of covariance matrix. \cr
  \code{LMM}    \tab Marginal test (LM(\eqn{\rho}|\eqn{\lambda}) or LM(\eqn{\lambda}|\eqn{\rho})) of spatial autocorrelation in the residuals. \cr
  }
}
\description{
Maximum likelihood estimation of the SUR model without spatial effects SUR-SIM, and  the spatial
SUR-SLX; SUR-SAR; SUR-SEM; SUR-SDM; SUR-SEDM and SUR-SARAR models
 (including possible restrictions of beta coefficients).
}
\details{
Maximum likelihood estimation of the models: \cr
\cr
SUR-SIM: SUR model without spatial effects\cr
\deqn{y_{gt} = y_{gt} + X_{gt} \beta_{g} + \epsilon_{gt}}
SUR-SLX: \cr
\deqn{y_{gt} = y_{gt} + X_{gt} \beta_{g} + WX_{gt} \theta_{g} + \epsilon_{gt}}
SUR-SAR: Spatial autorregresive model\cr
\deqn{y_{gt} = \lambda_{g} Wy_{gt} + X_{gt} \beta_{g} + \epsilon_{gt}}
SUR-SEM: Spatial error model\cr
\deqn{y_{gt} = X_{gt} \beta_{g} + u_{gt} ; u_{gt} = \rho_{g} Wu_{gt} + \epsilon_{gt}}
SUR-SDM: Spatial Durbin model\cr
\deqn{y_{gt} = \lambda_{g} Wy_{gt} + X_{gt} \beta_{g} + WX_{gt} \theta_{g} + \epsilon_{gt}}
SUR-SDEM: Spatial Durbin error model\cr
\deqn{y_{gt} = X_{gt} \beta_{g} + WX_{gt} \theta_{g} + u_{gt} ; u_{gt} = \rho_{g} W u_{gt} + \epsilon_{gt}}
SUR-SARAR: Spatial autoregressive model with spatial autoregressive error term\cr
\deqn{y_{gt} = \lambda_{g} Wy_{gt} + X_{gt} \beta_{g} + u_{gt} ; u_{gt} = \rho_{g} W u_{gt} + \epsilon_{gt}}
where y_{gt}, u_{gt} and \eqn{\epsilon_{gt}} are (nRx1) vectors; X_{gt} is a matrix of exogenous variables of
order (nRxsum(p)); \eqn{\lambda_g} and \eqn{\rho_g} are parametres of spatial dependence; W is a nRxnR matrix of spatial interactions.

By default an object create with \code{\link[Formula]{Formula}} is the input of this function. Alternatively a Y vector of dependent
variables and X matrix with independent varaibles could be include as imputs obtain by exaple with \code{\link{dgp_spSUR}}\cr
\cr
The marginal Multiplier tests are used to test for no correlation in one part of the model allowing for spatial correlation in the other. \cr
The LM(LM(\eqn{\rho}|\eqn{\lambda})) is the test for sustantive spatial autocorrelation in a model with spatial autocorrelation in error term.
The LM(\eqn{\lambda}|\eqn{\rho}) is the test for spatial error correlation in a model with sustanative spatial correlation.
}
\examples{

#################################################
######## CROSS SECTION DATA (nG>1; nT=1) ########
#################################################

#### Example 1: Spatial Phillips-Curve. Anselin (1988, p. 203)
## A SUR model without spatial effects
data(spc)
Tformula <- WAGE83 | WAGE81 ~ UN83 + NMR83 + SMSA | UN80 + NMR80 + SMSA
spcSUR.sim <-spsurml(Form=Tformula,data=spc,type="sim",W=Wspc)
summary(spcSUR.sim)

## A SUR-SLX model
spcSUR.slx <-spsurml(Form=Tformula,data=spc,type="slx",W=Wspc)
summary(spcSUR.slx)

## A SUR-SAR model
spcSUR.sar <-spsurml(Form=Tformula,data=spc,type="sar",W=Wspc)
summary(spcSUR.sar)

## A SUR-SEM model
spcSUR.sem <-spsurml(Form=Tformula,data=spc,type="sem",W=Wspc)
summary(spcSUR.sem)

## A SUR-SDM model
spcSUR.sdm <-spsurml(Form=Tformula,data=spc,type="sdm",W=Wspc)
summary(spcSUR.sdm)

## A SUR-SDEM model
spcSUR.sdem <-spsurml(Form=Tformula,data=spc,type="sdem",W=Wspc)
summary(spcSUR.sdem)

## A SUR-SARAR model
spcSUR.sarar <-spsurml(Form=Tformula,data=spc,type="sarar",W=Wspc)
summary(spcSUR.sarar)

#################################################
######## PANEL DATA (nG>1; nT>1)         ########
#################################################

#### Example 2: Homicides + Socio-Economics (1960-90)
Homicides and selected socio-economic characteristics for continental U.S. counties. Data for four decennial census years: 1960, 1970, 1980 and 1990.
https://geodacenter.github.io/data-and-lab/ncovr/

data(NAT)
Tformula <- HR80  | HR90 ~ PS80 + UE80 | PS90 + UE90
## A SUR-SIM model
NATSUR.sim <-spsurml(Form=Tformula,data=NAT,type="sim",W=W)
summary(NATSUR.sim)

## A SUR-SLX model
NATSUR.slx <-spsurml(Form=Tformula,data=NAT,type="slx",W=W)
summary(NATSUR.slx)

## A SUR-SAR model
NATSUR.sar <-spsurml(Form=Tformula,data=NAT,type="sar",W=W)
summary(NATSUR.sar)

## A SUR-SEM model
NATSUR.sem <-spsurml(Form=Tformula,data=NAT,type="sem",W=W)
summary(NATSUR.sem)

## A SUR-SDM model
NATSUR.sdm <-spsurml(Form=Tformula,data=NAT,type="sdm",W=W)
summary(NATSUR.sdm)

## A SUR-SDEM model
NATSUR.sdem <-spsurml(Form=Tformula,data=NAT,type="sdem",W=W)
summary(NATSUR.sdem)

#' ## A SUR-SARAR model
NATSUR.sarar <-spsurml(Form=Tformula,data=NAT,type="sarar",W=W)
summary(NATSUR.sarar)


##############################################
######## PANEL DATA (nG>1; nT>1) DEMEANING ###
##############################################

#### Example 3: Matrix as input argument
data(Sar)
nT <- 4 # Number of periods
nG <- 3 # Number equations
nR <- 49 # Number of spatial units
SUR.sar.ml <-spsurml(Y=Ysar,X=XXsar,nG=nG,nR=nR,nT=nT,p=5,W=Ws,type="sar")
summary(SUR.sar.ml)

#### Estimation with demeaning in nT (intercept of each equation dissapears)
SUR.sar.ml.dem <-spsurml(Y=Ysar,X=XXsar,nG=nG,nR=nR,nT=nT,p=5,W=Ws,type="sar",demean=TRUE)
summary(SUR.sar.ml.dem)

# Durbin case with demeaning in nT
SUR.sdm.ml.dem <-spsurml(Y=Ysar,X=XXsar,nG=nG,nR=nR,nT=nT,p=5,W=Ws,type="sdm",demean=TRUE)
summary(SUR.sdm.ml.dem)
}
\references{
Mur, J., Lopez, F., & Herrera, M. (2010). Testing for spatial effects in seemingly unrelated regressions. \emph{Spatial Economic Analysis}, 5(4), 399-440.
\cr
\cr
Lopez, F.A., Mur, J., & Angulo, A. (2014). Spatial model selection strategies in a SUR framework. The case of regional productivity in EU. \emph{Annals of Regional Science}, 53(1), 197-220.
\cr
\cr
Breusch T, Pagan A (1980) The Lagrange multiplier test and its applications to model specification in
 econometrics. \emph{Rev Econ Stud} 47: 239-254
}
\seealso{
\code{\link{lmtestspsur}}, \code{\link{spsurml}}, \code{\link{spsur3sls}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Jesús Mur  \tab \email{jmur@unizar.es} \cr
  }
}
